Create abstract super class for text like columns
